Refactored all DB code into storage package.
