+++ app/src/main/java/com/notemd/storage/AppDatabase.java
@@ -1,4 +1,4 @@
-package com.notemd;
+package com.notemd.storage;
 
 import android.content.Context;
